requirements.txt
aliyun_python_sdk_core==2.13.36
aliyun_python_sdk_ecs==4.24.82
apscheduler==3.11.1
asgiref==3.11.0
celery==5.6.0
channels==3.0.4
Django
django_fernet_fields_v2==0.9
kubernetes==34.1.0
openai==2.9.0
cryptography
paramiko
psutil==7.1.3
PyYAML==6.0.3
Requests==2.32.5
tencentcloud_sdk_python==3.1.7
whitenoise
openpyxl
pyte
redis
eventlet
crontab